requirements.txt
python-binance==0.3
pymongo==4.0.1
# pymongo[srv]==1.16.0
requests~=2.26.0
